# aidis-command/frontend/.openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AIDIS Command Backend API",
    "version": "1.0.0",
    "description": "RESTful API for AIDIS Command Backend with MCP Bridge Integration",
    "contact": {
      "name": "AIDIS Development Team",
      "email": "dev@aidis.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    },
    {
      "url": "http://localhost:3002/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiSuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "description": "Response data payload"
          },
          "correlationId": {
            "type": "string",
            "format": "uuid",
            "description": "Request correlation ID for tracking"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "version": {
                "type": "string",
                "example": "1.0.0"
              }
            }
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
      "BaseEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at"
        ]
      },
...
      "NamingEntry": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseEntity"
          },
          {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "variable",
                  "function",
                  "component",
                  "class",
                  "interface",
                  "module",
                  "file"
                ]
              },
              "context": {
                "type": "string"
              },
              "project_id": {
                "type": "string",
                "format": "uuid"
              },
              "project_name": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "active",
                  "deprecated",
                  "conflicted",
                  "pending"
                ]
              },
              "compliance_score": {
                "type": "number"
              },
              "usage_count": {
                "type": "number"
              },
              "created_by": {
                "type": "string"
              },
              "updated_by": {
                "type": "string"
              }
            },
            "required": [
              "name",
              "type",
              "status",
              "compliance_score",
              "usage_count"
            ]
          }
        ]
      },
      "NamingSearchResponse": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamingEntry"
            }
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          }
        },
        "required": [
          "entries",
          "total"
        ]
      },
      "RegisterNamingRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "type": {
            "type": "string",
            "enum": [
              "variable",
              "function",
              "component",
              "class",
              "interface",
              "module",
              "file"
            ]
          },
          "context": {
            "type": "string",
            "maxLength": 1000
          },
          "project_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "UpdateNamingRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "active",
              "deprecated",
              "conflicted",
              "pending"
            ]
          },
          "context": {
            "type": "string"
          }
        }
      },
      "NamingStats": {
        "type": "object",
        "properties": {
          "total_names": {
            "type": "integer"
          },
          "compliance": {
            "type": "number"
          },
          "deprecated": {
            "type": "integer"
          },
          "recent_activity": {
            "type": "integer"
          },
          "by_type": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "by_status": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "by_project": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "total_projects": {
            "type": "integer"
          }
        }
      },
      "NamingSuggestion": {
        "type": "object",
        "properties": {
          "suggested_name": {
            "type": "string"
          },
          "confidence": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          },
          "alternatives": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "suggested_name",
          "confidence"
        ]
      },
      "NamingAvailabilityResponse": {
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "conflicts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamingEntry"
            }
          }
        },
        "required": [
          "available"
        ]
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/naming": {
      "get": {
        "summary": "Search naming registry entries",
        "tags": [
          "Naming"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "variable",
                "function",
                "component",
                "class",
                "interface",
                "module",
                "file"
              ]
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "deprecated",
                "conflicted",
                "pending"
              ]
            }
          },
          {
            "in": "query",
            "name": "project_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "created_by",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entries retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/NamingSearchResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/naming/stats": {
      "get": {
        "summary": "Retrieve naming registry statistics",
        "tags": [
          "Naming"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics returned",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/NamingStats"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/naming/check/{name}": {
      "get": {
        "summary": "Check naming entry availability",
        "tags": [
          "Naming"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "variable",
                "function",
                "component",
                "class",
                "interface",
                "module",
                "file"
              ]
            },
            "description": "Optional type hint for the suggestion engine"
          }
        ],
        "responses": {
          "200": {
            "description": "Availability result",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/NamingAvailabilityResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/naming/suggest/{name}": {
      "get": {
        "summary": "Provide naming suggestions",
        "tags": [
          "Naming"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Suggestions returned",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/NamingSuggestion"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/naming/register": {
      "post": {
        "summary": "Register a new naming entry",
        "tags": [
          "Naming"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterNamingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Naming entry registered",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/NamingEntry"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/naming/{id}": {
      "get": {
        "summary": "Get naming entry by ID",
        "tags": [
          "Naming"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entry retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/NamingEntry"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update naming entry",
        "tags": [
          "Naming"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNamingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entry updated"
          }
        }
      },
      "delete": {
        "summary": "Delete naming entry",
        "tags": [
          "Naming"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entry deleted"
          }
        }
      }
    },
...
  }
}